Split out download and download status into its own class
