+++ filedistribution/src/main/java/com/yahoo/vespa/filedistribution/FileReferenceDownloader.java
@@ -12,11 +12,8 @@ import com.yahoo.vespa.config.ConnectionPool;
 import java.io.File;
 import java.time.Duration;
 import java.time.Instant;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
@@ -43,7 +40,7 @@ public class FileReferenceDownloader {
     /* Ongoing downloads */
     private final Downloads downloads = new Downloads();
     /* Status for ongoing and finished downloads */
-    private final DownloadStatuses downloadStatuses = new DownloadStatuses();
+    private final Downloads.DownloadStatuses downloadStatuses = new Downloads.DownloadStatuses();
     private final Duration downloadTimeout;
     private final Duration sleepBetweenRetries;
     private final Duration rpcTimeout;
@@ -98,7 +95,7 @@ public class FileReferenceDownloader {
     void completedDownloading(FileReference fileReference, File file) {
         Optional<FileReferenceDownload> download = downloads.get(fileReference);
         if (download.isPresent()) {
-            downloadStatuses.get(fileReference).ifPresent(DownloadStatus::finished);
+            downloadStatuses.get(fileReference).ifPresent(Downloads.DownloadStatus::finished);
             downloads.remove(fileReference);
             download.get().future().complete(Optional.of(file));
         } else {
@@ -157,7 +154,7 @@ public class FileReferenceDownloader {
 
     double downloadStatus(String file) {
         double status = 0.0;
-        Optional<DownloadStatus> downloadStatus = downloadStatuses.get(new FileReference(file));
+        Optional<Downloads.DownloadStatus> downloadStatus = downloadStatuses.get(new FileReference(file));
         if (downloadStatus.isPresent()) {
             status = downloadStatus.get().progress();
         }
@@ -165,7 +162,7 @@ public class FileReferenceDownloader {
     }
 
     void setDownloadStatus(FileReference fileReference, double completeness) {
-        Optional<DownloadStatus> downloadStatus = downloadStatuses.get(fileReference);
+        Optional<Downloads.DownloadStatus> downloadStatus = downloadStatuses.get(fileReference);
         if (downloadStatus.isPresent())
             downloadStatus.get().setProgress(completeness);
         else
@@ -173,7 +170,7 @@ public class FileReferenceDownloader {
     }
 
     Map<FileReference, Double> downloadStatus() {
-        return downloadStatuses.all().values().stream().collect(Collectors.toMap(DownloadStatus::fileReference, DownloadStatus::progress));
+        return downloadStatuses.all().values().stream().collect(Collectors.toMap(Downloads.DownloadStatus::fileReference, Downloads.DownloadStatus::progress));
     }
 
     public ConnectionPool connectionPool() {
@@ -187,86 +184,6 @@ public class FileReferenceDownloader {
         } catch (InterruptedException e) {
             Thread.interrupted(); // Ignore and continue shutdown.
         }
-    }
-
-    private static class Downloads {
-        private final Map<FileReference, FileReferenceDownload> downloads = new ConcurrentHashMap<>();
-
-        void add(FileReferenceDownload fileReferenceDownload) {
-            downloads.put(fileReferenceDownload.fileReference(), fileReferenceDownload);
-        }
-
-        void remove(FileReference fileReference) {
-            downloads.remove(fileReference);
-        }
-
-        Optional<FileReferenceDownload> get(FileReference fileReference) {
-            return Optional.ofNullable(downloads.get(fileReference));
-        }
-    }
-
-    private static class DownloadStatus {
-        private final FileReference fileReference;
-        private double progress; // between 0 and 1
-        private final Instant created;
-
-        DownloadStatus(FileReference fileReference) {
-            this.fileReference = fileReference;
-            this.progress = 0.0;
-            this.created = Instant.now();
-        }
-
-        public FileReference fileReference() {
-            return fileReference;
-        }
-
-        public double progress() {
-            return progress;
-        }
-
-        public void setProgress(double progress) {
-            this.progress = progress;
-        }
-
-        public void finished() {
-            setProgress(1.0);
-        }
-
-        public Instant created() {
-            return created;
-        }
-    }
-
-    /* Status for ongoing and completed downloads, keeps at most status for 100 last downloads */
-    private static class DownloadStatuses {
-
-        private static final int maxEntries = 100;
-
-        private final Map<FileReference, DownloadStatus> downloadStatus = new ConcurrentHashMap<>();
-
-        void add(FileReference fileReference) {
-            add(fileReference, 0.0);
-        }
-
-        void add(FileReference fileReference, double progress) {
-            DownloadStatus ds = new DownloadStatus(fileReference);
-            ds.setProgress(progress);
-            downloadStatus.put(fileReference, ds);
-            if (downloadStatus.size() > maxEntries) {
-                Map.Entry<FileReference, DownloadStatus> oldest =
-                        Collections.min(downloadStatus.entrySet(), Comparator.comparing(e -> e.getValue().created));
-                downloadStatus.remove(oldest.getKey());
-            }
-        }
-
-        Optional<DownloadStatus> get(FileReference fileReference) {
-            return Optional.ofNullable(downloadStatus.get(fileReference));
-        }
-
-        Map<FileReference, DownloadStatus> all() {
-            return Map.copyOf(downloadStatus);
-        }
-
     }
 
 }
